# api/openapi_new.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portcullis",
    "version": "0.8.0"
  },
  "paths": {
...
    "/auth/sso/{slug}/login": {
      "get": {
        "tags": [
          "sso"
        ],
        "summary": "Sso Login",
        "description": "Start the OIDC authorization-code flow (public redirect endpoint).",
        "operationId": "sso_login_auth_sso__slug__login_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sso/{slug}/callback": {
      "get": {
        "tags": [
          "sso"
        ],
        "summary": "Sso Callback",
        "description": "Complete the OIDC flow and redirect the browser with a one-time token.",
        "operationId": "sso_callback_auth_sso__slug__callback_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/telemetry/heartbeat": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Heartbeat",
        "description": "Record (or refresh) an anonymous self-host install heartbeat.",
        "operationId": "heartbeat_v1_telemetry_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/telemetry/summary": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Admin Telemetry Summary",
        "description": "Return aggregate install counts (platform admin only).",
        "operationId": "admin_telemetry_summary_admin_telemetry_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceSummaryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/telemetry/instances": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Admin Telemetry Instances",
        "description": "Return all known self-host installs, most recently seen first.",
        "operationId": "admin_telemetry_instances_admin_telemetry_instances_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstanceView"
                  },
                  "title": "Response Admin Telemetry Instances Admin Telemetry Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstanceHeartbeatRequest": {
        "properties": {
          "install_id": {
            "type": "string",
            "format": "uuid",
            "title": "Install Id"
          },
          "version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Version"
          },
          "server_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Server Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "install_id",
          "version"
        ],
        "title": "InstanceHeartbeatRequest",
        "description": "Anonymous self-host heartbeat payload.\n\n``install_id`` is the install's persistent identifier (generated on first\nboot and kept in its local config); ``version`` is the running Portcullis\nversion and ``server_count`` the number of registered MCP servers.  The\npayload deliberately contains no tenant or user data."
      },
      "InstanceSummaryView": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "active_24h": {
            "type": "integer",
            "title": "Active 24H"
          }
        },
        "type": "object",
        "required": [
          "total",
          "active_24h"
        ],
        "title": "InstanceSummaryView",
        "description": "Aggregate install counts for the Phase 4 self-host funnel review."
      },
      "InstanceView": {
        "properties": {
          "install_id": {
            "type": "string",
            "format": "uuid",
            "title": "Install Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "server_count": {
            "type": "integer",
            "title": "Server Count"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "install_id",
          "version",
          "server_count",
          "first_seen_at",
          "last_seen_at"
        ],
        "title": "InstanceView",
        "description": "Safe response schema for a known self-host install."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}